requirements.txt
Django==1.10
psycopg2==2.6.2
beautifulsoup4
requests
pytz
django-bootstrap

# django-rest
djangorestframework
markdown
django-filter
djangorestframework-gis




# Test project
pep8
pylint
mock
coverage
